pyproject.toml
[tool.poetry]
name = "fixafel"
version = "0.1.0"
description = "fix tool"
authors = ["Hector Varas <hvarass@falabella.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
google-auth = "^2.30.0"
google-cloud-firestore = "^2.16.0"
dependency-injector="4.42.0b1"
pandas = "^2.2.2"
tabulate = "^0.9.0"
ipython = "^8.25.0"
containers = "^0.0.4"
providers = "^0.0.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
